# apps/web/public/openapi.yaml
openapi: 3.0.3
info:
  title: TimeTiles API
  version: 1.0.0
  description: >-
    API for querying and aggregating geospatial event data in TimeTiles.


    ## Overview


    TimeTiles provides endpoints for:

    - **Event listing** with pagination and filtering

    - **Aggregations** by catalog or dataset

    - **Temporal histograms** for time-based analysis

    - **Map clustering** for efficient visualization of large datasets

    - **Cluster statistics** for consistent visualization scaling


    ## Authentication


    Most endpoints support optional authentication. Authenticated users may have
    access to additional catalogs and datasets.


    ## Filtering


    All event endpoints support common filter parameters:

    - `catalog`: Filter by catalog ID

    - `datasets`: Filter by dataset IDs (comma-separated or multiple params)

    - `startDate` / `endDate`: Temporal filtering (ISO 8601 date format)

    - `bounds`: Geographic bounding box as JSON string
  contact:
    name: TimeTiles
    url: https://github.com/timetiles/timetiles
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /
    description: Current server
...
components:
  schemas:
    EventItem:
      type: object
      properties:
        id:
          type: number
        dataset:
          type: object
          properties:
            id:
              type: number
            title:
              type: string
            catalog:
              type: string
          required:
            - id
        data:
          type: object
          additionalProperties:
            nullable: true
        location:
          type: object
          nullable: true
          properties:
            longitude:
              type: number
            latitude:
              type: number
          required:
            - longitude
            - latitude
        eventTimestamp:
          type: string
        isValid:
          type: boolean
      required:
        - id
        - dataset
        - data
        - location
        - eventTimestamp
        - isValid
    EventListResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/EventItem"
        pagination:
          type: object
          properties:
            page:
              type: number
            limit:
              type: number
            totalDocs:
              type: number
            totalPages:
              type: number
            hasNextPage:
              type: boolean
            hasPrevPage:
              type: boolean
            nextPage:
              type: number
              nullable: true
            prevPage:
              type: number
              nullable: true
          required:
            - page
            - limit
            - totalDocs
            - totalPages
            - hasNextPage
            - hasPrevPage
            - nextPage
            - prevPage
      required:
        - events
        - pagination
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          nullable: true
      required:
        - error
    AggregationItem:
      type: object
      properties:
        id:
          anyOf:
            - type: number
            - type: string
        name:
          type: string
        count:
          type: integer
      required:
        - id
        - name
        - count
    AggregateResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AggregationItem"
        total:
          type: integer
        groupedBy:
          type: string
      required:
        - items
        - total
        - groupedBy
...
paths:
  /api/v1/events:
    get:
      tags:
        - Events
      summary: List events with pagination
      description: Returns a paginated list of events matching the specified filters.
        Events are returned with enriched data including extracted title,
        description, and location.
      parameters:
        - schema:
            type: integer
            nullable: true
          required: false
          name: catalog
          in: query
        - schema:
            type: array
            items: &a1
              type: integer
          required: false
          name: datasets
          in: query
        - schema:
            type: string
            format: date
          required: false
          name: startDate
          in: query
        - schema:
            type: string
            format: date
          required: false
          name: endDate
          in: query
        - schema:
            type: string
          required: false
          name: bounds
          in: query
        - schema:
            type: integer
            minimum: 1
            default: 1
          required: false
          name: page
          in: query
        - schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
          required: false
          name: limit
          in: query
        - schema:
            type: string
            default: -eventTimestamp
          required: false
          name: sort
          in: query
      responses:
        "200":
          description: Paginated event list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/events/stats:
    get:
      tags:
        - Events
      summary: Aggregate event counts by catalog or dataset
      description: Returns event counts grouped by catalog or dataset. When datasets
        are explicitly filtered, all selected datasets appear in results (with 0
        count if no events match in the current viewport).
      parameters:
        - schema:
            type: integer
            nullable: true
          required: false
          name: catalog
          in: query
        - schema:
            type: array
            items: *a1
          required: false
          name: datasets
          in: query
        - schema:
            type: string
            format: date
          required: false
          name: startDate
          in: query
        - schema:
            type: string
            format: date
          required: false
          name: endDate
          in: query
        - schema:
            type: string
          required: false
          name: bounds
          in: query
        - schema:
            type: string
            enum:
              - catalog
              - dataset
          required: true
          name: groupBy
          in: query
      responses:
        "200":
          description: Aggregated event counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AggregateResponse"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...